+++ CHANGELOG.md
@@ -12,5 +12,5 @@ We follow [Semantic Versions](https://semver.org/).
 
 - Initial release
 - Use double quotes for string literals
-- Set up CLI with Click
-- Implement "--version" CLI option to show the current version
+- Set up CLI with `click`
+- Implement `--version` CLI option to show the current version