+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 We follow [Semantic Versions](https://semver.org/).
 
 
-## Version 0.1.1 (unreleased)
+## Version 0.1.1
 
 - Configure logging with `structlog`
 - Implement `--debug` CLI option (sets the logging level to `DEBUG`)